Silence compile warnings on Irix.

+++ src/access_log.c
@@ -787,7 +787,7 @@ accessLogGetNewLogFormatToken(logformat_token * lt, char *def, enum log_quote *q
 	lt->zero = 1;
 	cur++;
     }
-    if (isdigit(*cur))
+    if (isdigit((int) *cur))
 	lt->width = strtol(cur, &cur, 10);
     if (*cur == '.')
 	lt->precision = strtol(cur + 1, &cur, 10);
